static path fixes
