+++ network-api/networkapi/wagtail_l10n_customization/static/js/language_toggles.js
